Refactor code into different files using the MVC pattern
